square root

+++ square_root/include/square_root/square_root.hpp
@@ -7,7 +7,7 @@ class SquareRoot {
 
     ~SquareRoot();
 
-    int getSquareRoot(int num);
+    double getSquareRoot(double num);
 };
